Move the noises to the simulation module
Import NoiseModel as noises.NoiseModel to avoid a weird circular import
