+++ pulser/sampler/sampler.py
@@ -10,10 +10,9 @@ from typing import Callable, List, Optional, cast
 
 import numpy as np
 
-import pulser.sampler.noises as noises
+import pulser.simulation.noises as noises
 from pulser.channels import Channel
 from pulser.pulse import Pulse
-from pulser.sampler.noises import NoiseModel
 from pulser.sampler.samples import QubitSamples
 from pulser.sequence import Sequence, _TimeSlot
 
@@ -21,8 +20,8 @@ from pulser.sequence import Sequence, _TimeSlot
 def sample(
     seq: Sequence,
     modulation: bool = False,
-    common_noises: Optional[list[NoiseModel]] = None,
-    global_noises: Optional[list[NoiseModel]] = None,
+    common_noises: Optional[list[noises.NoiseModel]] = None,
+    global_noises: Optional[list[noises.NoiseModel]] = None,
 ) -> dict:
     """Samples the given Sequence and returns a nested dictionary.
